Resolve use-after-free in TKrnlLList
When orphaning the last element in a list, adding another later,
then iterating over the result, somehow the iterator was going past
the end of the list and getting nonsense pointers.  I can't really
see how this could possibly happen from this bug, that was only
writing to a structure that no longer existed, never reading from
it, as far as I can tell.  But that's undefined behaviour for ya.