Re-implement time-zone func.s using tzset
The existing implementation using gettimeofday was not working for
me, returning 0 as both the offset and the name.  The manual goes
so far as to say that even just passing a non-null pointer to
the timezone argument of gettimeofday causes unspecified behaviour!

I'm not sure how this new implementation will fare with daylight
savings time; I don't have a 32-bit faketime to try it out.
But, it's less than a week away, so I'll just test it then.